remove the function test in another page
