+++ src/server.js
@@ -3,27 +3,10 @@ var fs = require('fs');
 
 var wordPackage = require('word-list-json');
 
+var helpers = require('./helpers');
+
 var NUM_RESULTS = 10;
 
-function getWords(searchTerm, wordList, maxResults) {
-  var newArr = []
-  if (searchTerm === ''){
-    return newArr;
-  }
-  var searchTermLowerCase = searchTerm.toLowerCase();
-
-  for (var i = 0, n = wordList.length; i < n; i++) {
-    var word = wordList[i];
-    if (maxResults && newArr.length === maxResults) {
-      break;
-    }
-    if (word.startsWith(searchTermLowerCase)){
-        newArr.push(word);
-    }
-  };
-
-  return newArr;
-}
 
 function handler(request, response) {
   var url = request.url;
@@ -72,7 +55,7 @@ function handler(request, response) {
     // we want the value after equals sign
     var searchTerm = url.split('=')[1];
     // the result of the client search (its an array of words)
-    var wordArray = getWords(searchTerm, wordPackage, NUM_RESULTS);
+    var wordArray = helpers.getWords(searchTerm, wordPackage, NUM_RESULTS);
     // we send the response
     response.writeHead(200, {'Content-Type':'text/plain'});
     response.end(JSON.stringify(wordArray));
