+++ tests/tests.js
@@ -1,24 +1,7 @@
 var test = require('tape');
 
-function getWords(searchTerm, wordList, maxResults) {
-  var newArr = []
-  if (searchTerm === ''){
-    return newArr;
-  }
-  searchTerm = searchTerm.toLowerCase();
+var helpers = require('../src/helpers.js');
 
-  for (var i = 0, n = wordList.length; i < n; i++) {
-    var word = wordList[i];
-    if (maxResults && newArr.length === maxResults) {
-      break;
-    }
-    if (word.startsWith(searchTerm)){
-        newArr.push(word);
-    }
-  };
-
-  return newArr;
-}
 
 test('Check tape is working with a simple passing test', function(t) {
   t.pass('a message to print on succes');
@@ -29,7 +12,7 @@ test('check get the words started in letter hhh',function(t){
   wordList = ["animal","bat","hello"];
   searchTerm = "hhh";
   t.deepEqual(
-    getWords(searchTerm, wordList),
+    helpers.getWords(searchTerm, wordList),
     [],
     'should return []'
   );
@@ -40,7 +23,7 @@ test('check get the words started in letter a',function(t){
   wordList = ["animal","bat","hello"];
   searchTerm = "a";
   t.deepEqual(
-    getWords(searchTerm, wordList, 10),
+    helpers.getWords(searchTerm, wordList, 10),
     ['animal'],
     'should return ["animal"]'
   );
@@ -51,7 +34,7 @@ test('check get the words started in letter a',function(t){
   wordList = ["animal","apple","hello"];
   searchTerm = "a";
   t.deepEqual(
-    getWords(searchTerm, wordList, 10),
+    helpers.getWords(searchTerm, wordList, 10),
     ['animal','apple'],
     'should return ["animal","apple"]'
   );
@@ -62,7 +45,7 @@ test('check get the words started in letter b',function(t){
   wordList = ["animal","apple","hello","banana"];
   searchTerm = "b";
   t.deepEqual(
-    getWords(searchTerm, wordList, 10),
+    helpers.getWords(searchTerm, wordList, 10),
     ['banana'],
     'should return ["banana"]'
   );
@@ -73,7 +56,7 @@ test('check get the words started in letter B',function(t){
   wordList = ["animal","apple","hello","banana"];
   searchTerm = "B";
   t.deepEqual(
-    getWords(searchTerm, wordList, 10),
+    helpers.getWords(searchTerm, wordList, 10),
     ['banana'],
     'should return ["banana"]'
   );
@@ -84,7 +67,7 @@ test('check the function returns no words when the input is an empty string',fun
   wordList = ["animal","apple","hello","banana"];
   searchTerm = "";
   t.deepEqual(
-    getWords(searchTerm, wordList, 10),
+    helpers.getWords(searchTerm, wordList, 10),
     [],
     'should return []'
   );
@@ -95,7 +78,7 @@ test('check the function returns maximum 2 words',function(t){
   wordList = ["animal","apple","hello","banana","arabic"];
   searchTerm = "a";
   t.deepEqual(
-    getWords(searchTerm, wordList ,2),
+    helpers.getWords(searchTerm, wordList ,2),
     ["animal","apple"],
     'should return ["animal","apple"]'
   );
